backend/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.4.4
defusedxml==0.6.0
Django==3.1.6
django-allauth==0.44.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.2
idna==2.10
Markdown==3.3.3
oauthlib==3.1.0
Pillow==8.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
